requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
black==24.10.0
    # via
    #   pipa (pyproject.toml)
    #   flake8-black
build==1.3.0
    # via pip-tools
cfgv==3.4.0
    # via pre-commit
click==8.3.0
    # via
    #   pipa (pyproject.toml)
    #   black
    #   pip-tools
coverage==7.11.0
    # via pytest-cov
distlib==0.4.0
    # via virtualenv
filelock==3.20.0
    # via virtualenv
flake8==7.3.0
    # via
    #   pipa (pyproject.toml)
    #   flake8-black
flake8-black==0.4.0
    # via pipa (pyproject.toml)
identify==2.6.15
    # via pre-commit
iniconfig==2.3.0
    # via pytest
isort==5.13.2
    # via pipa (pyproject.toml)
jinja2==3.1.6
    # via pipa (pyproject.toml)
joblib==1.5.2
    # via scikit-learn
kneed==0.8.5
    # via pipa (pyproject.toml)
markdown-it-py==3.0.0
    # via pipa (pyproject.toml)
markupsafe==3.0.3
    # via jinja2
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.1.0
    # via black
nodeenv==1.9.1
    # via pre-commit
numpy==1.26.4
    # via
    #   pipa (pyproject.toml)
    #   kneed
    #   pandas
    #   scikit-learn
    #   scipy
packaging==25.0
    # via
    #   black
    #   build
    #   plotly
    #   pytest
pandas==2.2.3
    # via pipa (pyproject.toml)
pathspec==0.12.1
    # via black
pip==25.3
    # via pip-tools
pip-tools==7.5.1
    # via pipa (pyproject.toml)
platformdirs==4.5.0
    # via
    #   black
    #   virtualenv
plotly==5.24.1
    # via pipa (pyproject.toml)
pluggy==1.6.0
    # via pytest
pre-commit==3.8.0
    # via pipa (pyproject.toml)
pycodestyle==2.14.0
    # via flake8
pyflakes==3.4.0
    # via flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==7.4.4
    # via
    #   pipa (pyproject.toml)
    #   pytest-cov
pytest-cov==5.0.0
    # via pipa (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   pipa (pyproject.toml)
    #   pre-commit
scikit-learn==1.7.2
    # via pipa (pyproject.toml)
scipy==1.16.3
    # via
    #   kneed
    #   scikit-learn
setuptools==80.9.0
    # via pip-tools
six==1.17.0
    # via python-dateutil
tabulate==0.9.0
    # via pipa (pyproject.toml)
tenacity==9.1.2
    # via plotly
threadpoolctl==3.6.0
    # via scikit-learn
tzdata==2025.2
    # via pandas
uv==0.9.10
    # via pipa (pyproject.toml)
virtualenv==20.35.4
    # via pre-commit
wheel==0.45.1
    # via pip-tools
